Move presentational components to components/
