+++ packages/sharing-editor/src/Editor/components/FileUploader.module.scss
