+++ packages/sharing-editor/src/Editor/components/FileUploader.tsx
@@ -1,0 +1,15 @@
+import React from "react";
+import { RiUpload2Line } from "react-icons/ri";
+import styles from "./FileUploader.module.scss";
+
+type InputProps = JSX.IntrinsicElements["input"];
+function FileUploader(props: InputProps) {
+  return (
+    <label className={styles.label}>
+      <RiUpload2Line />
+      <input type="file" {...props} className={styles.fileInput} />
+    </label>
+  );
+}
+
+export default FileUploader;
