+++ packages/sharing-editor/src/Editor/components/Tab.module.scss
