+++ packages/sharing-editor/src/Editor/components/Tab.tsx
