+++ packages/sharing-editor/src/Editor/components/TabBar.module.scss
