+++ packages/sharing-editor/src/Editor/components/TabBar.tsx
