+++ packages/sharing-editor/src/Editor/components/Toolbar.module.scss
