+++ packages/sharing-editor/src/Editor/components/variables.scss
